docs: Fix pitito.Field docstring
Returns added as was automatically added in the wrong location and caused a warning.

Moved docstring to the helper class as this is how sphinx obtains the docstring for partial functions; neatens code as the docstring class and assignment is no longer necessary.

Corrected out of date Pydantic docs link, added api and guidance links, added link to return type. Reworded slightly.